hibernate setup libraries and configuration files
