--- pom.xml
@@ -1,25 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.ProjectNarwhal</groupId>
-    <artifactId>ProjectNarwhal</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-    <name>First Hibernate Tutorial</name>
-
-    <build>
-         <!-- we dont want the version to be part of the generated war file name -->
-         <finalName>${artifactId}</finalName>
-    </build>
-
-    <dependencies>
-        <dependency>
-		    <groupId>org.hibernate</groupId>
-	    	<artifactId>hibernate-core</artifactId>
-	    	<version>3.6.3.Final</version>
-		</dependency>
-    </dependencies>
-
-</project>
