+++ src/hibernate.cfg.xml
@@ -1,0 +1,38 @@
+<!DOCTYPE hibernate-configuration PUBLIC
+        "-//Hibernate/Hibernate Configuration DTD 3.0//EN"
+        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
+
+<hibernate-configuration>
+	<session-factory>
+	
+		<property name="hibernate.connection.driver_class">com.microsoft.sqlserver.jdbc.SQLServerDriver</property>
+	    <property name="hibernate.connection.url">com.microsoft.sqlserver.jdbc.SQLServerDriver</property>
+	    <property name="hibernate.connection.username">sa</property>
+	    <property name="hibernate.connection.password">lal</property>
+	    <property name="dialect">org.hibernate.dialect.SQLServerDialect</property>
+	
+	    <property name="hibernate.connection.pool_size">100</property>        
+	    <property name="show_sql">true</property>
+	
+	    <!-- Mapping files -->
+	    <mapping resource="model.hbm.xml"/>
+	
+	   <!-- Database connection settings --
+	   <property name="hibernate.connection.driver_class">oracle.jdbc.driver.OracleDriver</property>
+	   <property name="hibernate.connection.username">system</property>
+	   <property name="hibernate.connection.password">admin</property>
+	   <property name="hibernate.connection.url">jdbc:oracle:thin:@127.0.0.1:1521:XE</property>
+	
+	   <!-- SQL dialect --
+	   <property name="hibernate.dialect">org.hibernate.dialect.Oracle10gDialect</property>
+	
+	   <!-- Echo all executed SQL to sysout --
+	   <property name="show_sql">true</property>
+	
+	   <!-- Create/Update the database schema on startup --
+	   <property name="hibernate.hbm2ddl.auto">update</property>
+	   <!-- Mapping file --
+	   <mapping resource="Student.hbm.xml" />-->
+	
+	</session-factory>
+</hibernate-configuration>
